test/integration: Run test cases in parallel
